+++ inputs/base.js
@@ -1,0 +1,20 @@
+const baseTemplateStyle = `
+<style>
+    :focus {
+        outline: none;
+    }
+</style>
+`;
+
+export class Base extends HTMLElement {
+    constructor() {
+        super();
+
+        this.attachShadow({ mode: "open" });
+        this.shadowRoot.innerHTML += baseTemplateStyle;
+
+        this.shadowRoot.addEventListener("change", (e) => {
+            this.dispatchEvent(new Event("change", { bubbles: true }));
+        });
+    }
+}
