+++ jsconfig.json
@@ -1,5 +1,6 @@
 {
-  "compilerOptions": {
-    "checkJs": true
-  }
+    "compilerOptions": {
+        "module": "ES6",
+        "checkJs": true
+    }
 }
